Mo More Casting When Initializing I2C

+++ examples/RPi_4/wiringPi/example_rpi4_wiringPi.cpp
@@ -17,7 +17,7 @@ int main()
 {
   I2C_Interface* i2c_if = new WIRINGPI_I2C_IF();
   /* initialize fast I2C 400 kHz */
-  if(i2c_if->Init_I2C((uint8_t)MPU6050_ADDRESS) != I2C_STATUS_SUCCESS)
+  if(i2c_if->Init_I2C(MPU6050_ADDRESS) != I2C_STATUS_SUCCESS)
   {
     std::cout << "I2C initialization failed!\n";
     delete i2c_if;
